+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## XX.X.X
-* Added MetricHelper class and "overriddenMetrics" field in the configuration object, allowing to pass custom or overridden metrics.
+* Added functionality to allow passing custom or overridden metrics.
 
 ## 23.06.1
 * Added app version metric to every request sent.